Gen 1 personal tables

+++ src/personal_info/traits/base_stat.rs
@@ -1,18 +1,23 @@
 use crate::{PKError, PKResult};
 
 pub trait BaseStat {
-    fn hp(&self) -> u16;
-    fn atk(&self) -> u16;
-    fn def(&self) -> u16;
-    fn spa(&self) -> u16;
-    fn spd(&self) -> u16;
-    fn spe(&self) -> u16;
+    fn hp(&self) -> u8;
+    fn atk(&self) -> u8;
+    fn def(&self) -> u8;
+    fn spa(&self) -> u8;
+    fn spd(&self) -> u8;
+    fn spe(&self) -> u8;
 
     fn get_base_state_total(&self) -> u16 {
-        self.hp() + self.atk() + self.def() + self.spa() + self.spd() + self.spe()
+        self.hp() as u16
+            + self.atk() as u16
+            + self.def() as u16
+            + self.spa() as u16
+            + self.spd() as u16
+            + self.spe() as u16
     }
 
-    fn get_base_state_value(&self, index: usize) -> PKResult<u16> {
+    fn get_base_state_value(&self, index: usize) -> PKResult<u8> {
         match index {
             0 => Ok(self.hp()),
             1 => Ok(self.atk()),
@@ -24,7 +29,7 @@ pub trait BaseStat {
         }
     }
 
-    fn get_sorted_stat_indices(&self, stats: &mut [(usize, u16)]) -> PKResult<()> {
+    fn get_sorted_stat_indices(&self, stats: &mut [(usize, u8)]) -> PKResult<()> {
         for (i, stat) in stats.iter_mut().enumerate() {
             *stat = (i, self.get_base_state_value(i)?);
         }
